### 2021-08-06 #3578 @whedon
[REVIEW]: The Neuroimaging Data Model (NIDM) Linear Regression Tool
**Submitting author:** @AKUMAR0019 (<a href="http://orcid.org/0000-0001-8076-7620">Ashmita Kumar</a>)
**Repository:** <a href="https://github.com/incf-nidash/PyNIDM" target ="_blank">https://github.com/incf-nidash/PyNIDM</a>
**Version:** v.3.8.2
