### 2021-08-20 #3578 @robbisg
Hi @kthyng , sorry for being unclear.
This paper is about a linear regression tool which is only a portion of the main package. The repository link (https://github.com/AKUMAR0019/PyNIDM) is a fork of the main repository. Unfortunately, this fork doesn't contain the linear regression tool but only the JOSS paper files, so it is difficult to check whether the tool has tests, enough lines of code and so on. If I did something uncorrectly, I apologize.
