### 2021-08-20 #3578 @osorensen
Thanks for clarifying @robbisg. If I understand you correctly, the overview of contributors given in the source repository (https://github.com/AKUMAR0019/PyNIDM/graphs/contributors) shows all contributions to a larger package, and I understand that this makes it hard to assess the point on the review checklist regarding Contribution and authorship.

@AKUMAR0019, could you please clarify this issue. Are the contributions to the repository https://github.com/incf-nidash/PyNIDM directly related to the development of the software described in this submission?
