### 2021-08-21 #3578 @AKUMAR0019
Hi @robbisg and @osorensen. Thank you for your review! Please see my point by point responses below.

1. I worked on this project in a fork of the main PyNIDM repository and once the code functioned correctly it was merged into the main pynidm repository via a pull request. The main repository does contain the regression tool, found at this location: https://github.com/incf-nidash/PyNIDM/tree/master/nidm/experiment/tools.  All PyNIDM tools are contained here and integrated with the overall project using the Click python library.  This means anyone who installs the PyNIDM toolbox gets access to all of the tools that are supported for NIDM via the command line `pynidm [tool]`.  With respect to the linear regression too, it is run using `pynidm linear-regression`.   The commit history of the tool can be traced in my fork of pynidm (https://github.com/AKUMAR0019/PyNIDM ).

2. The PyNIDM linreg tool is related to the whole toolbox found in PyNIDM. It uses functions in the PyNIDM library, specifically the RestParser to query the data the RestParser is over 700 lines of code as is nidm_linreg.py so in total the tool is approximately 1400 lines of code. Since the linear regression tool works on the graphs supported by this package, the contributions in the repository are directly related to the development of the software. It is one of the many tools supporting NIDM graphs, all of which are included in this integrated PyNIDM toolbox.

3. Thank you for your suggestion to improve our description of the command line parameters and how to format the modal definitions.  We are working on revising this section of the manuscript.

4. Thank you for your suggestion to better describe the novelty of the tool.  We are working on revising this text in the manuscript.

5. The PyNIDM toolbox has lost of test cases. We are currently integrating the test cases for the linear regression tool and get back to you with a link to those specific to the linear regression tool.  The test cases for the RestParser (a functional component of the linear regression tool) can be found here: https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/tests/test_rest.py
