### 2021-08-21 #3578 @osorensen
Thanks for clarifying, @AKUMAR0019.

Regarding point 1, I see that the fork reports that it's 49 commits ahead of master. Does this mean that you are still developing the code in this branch, and create pull requests onto the repository incf-nidash/PyNIDM when done?

Second, and this may be a question to @kthyng or @openjournals/joss-eics, since the final package will be available in the repository incf-nidash/PyNIDM, I wonder if we should eventually set the repository for this submission to incf-nidash/PyNIDM, rather than the fork AKUMAR0019/PyNIDM?
