### 2021-08-21 #3578 @robbisg
Thanks for clarifying @AKUMAR0019.

I have a suggestion given your explanation of point 2. To my point of view, I think that a more detailed description of the input structure can be very helpful to clarify the tool, for example you mentioned that linear regression is performed on graph, so what are these graphs representing and which information is coded in these graphs?
