### 2021-08-22 #3578 @AKUMAR0019
Thank you for your responses @osorensen and @robbisg. Please see my point by point responses below.

1. @osorensen The 49 commits are associated with a different tool I am completing that is separate from nidm_linreg.py. In other words, those commits are not related to the paper. I will eventually create pull requests onto the repository incf-nidash/PyNIDM when done, but the complete linear regression tool is already in the main repository incf-nidash/PyNIDM.

2. @robbisg Thank you for your suggestion to include a more detailed description of the graphs and what is coded in them. We are working on revising this part in the manuscript.
