### 2021-08-23 #3578 @osorensen
Thanks @AKUMAR0019, that sounds absolutely fine regarding point 1. However, if repository incf-nidash/PyNIDM is the one where users can contribute to the code, we should reference that repository in the paper (if accepted) rather than AKUMAR0019/PyNIDM which is currently stated as the repository for this submission. Does that make sense?
