### 2021-08-25 #3578 @osorensen
Sorry about that @htwangtw, but good to hear that you're back at work. It should be possible to turn on notifications for this issue by clicking the "Notications" button on the right hand side if you scroll up to the top of the page.
