### 2021-08-30 #3578 @htwangtw
Hi all,

Thanks for writing this tool and invite me to review! This is my first JOSS review so please let me know

Since the submission refers to the linear regression tool in the pyNIDM package, I am not sure what is the most appropriate way to address some of the points. If anything is out of scope, please feel free to point it out @osorensen.

## Paper

The manuscript explains the purpose of the linear regression tool in PyNIDM with good clarity and examples. The software provides an unique application on NIDM.

More descriptions on the NIDM graph would be beneficial to the general reader. Ideally, the paper should be a self-contained document for the user to understand the software and usage without researching on other source materials.

Typesetting of the reference is not consistent with JOSS guideline, such as in Line 16, the correct way to type set it is  (NIDM; Keator et al., 2013; NIDM Working Group; Maumet et al., 2016). The markdown syntax is `[@author1:2001; @author2:2001]`. Please see [JOSS example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

## Software

The software is difficult to review as there are some issues that are going to impact the whole repository, not just the function added itself.

### Issues related to the linear regression module:

 - In the documentation, there should be a minimal explanation of NIDM graph.
 - The documentation should specify the syntax used for regression model. It is not standard `patsy` syntax used in `statsmodels`, from what I can see.
 - There's no/incomplete docstring for most of the functions. Adding docs will help future maintainer to understand the functions.
 - I cannot find automated test for the linear regression functions. In fact the most recent automated test was ran 3 months ago from what I can see.
 - In the header of the doc, the license state is GPL but the repo is Apache 2.0. Also GitHub doesn't seem to pick up the license type. Please double check this.

Here are two potentially out-of-scope issues. I am happy to have the editors decide if this is too much of a ask:
 - Avoid global variables. Passing variables to the respective functions will clarify the dependency of between inputs and outputs. This will make the code base more maintainable and easier to write tests.
 - There's no check on the validity of user inputs. Raising warning for invalid input can reduce the burden on users.
