### 2021-09-06 #3578 @AKUMAR0019
Thanks for your review @htwangtw. Please see my point by point responses below:

1. Thank you for your suggestion to improve our description of the command line parameter syntax and how to format the modal definitions. We are working on revising this section of the manuscript, as well as the documentation. These are being reviewed by the team right now.
2. Thank you for your suggestion to better describe NIDM. We are working on revising this text in the manuscript.
3. Thank you for your feedback on the references section. We will work on fixing that.
4. The PyNIDM toolbox has lots of test cases. We are currently integrating the test cases for the linear regression tool and get back to you with a link to those specific to the linear regression tool. The test cases for the RestParser (a functional component of the linear regression tool) can be found here: https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/tests/test_rest.py
5. We will check on the issue with the license and get back to you.
6. Currently, the code does check input validity and raises errors, which we did in order to reduce the burden on the users. The code handles erroneous inputs with custom error messages. For example, if a variable is not found, instead of letting it return a convoluted error message, we have it specify which variables were not found in which files, so the user has better feedback on what to change in the next run of the command. Error checks also exist for improper syntax and for cases where there are less than 20 points of data. Once these error messages are given, the code is stopped. However, we would appreciate more feedback on other parts of the user input to check.
7. Thank you for your feedback on global variables. We will do our best to change the code to use parameters instead.
8. Thank you for your feedback on docstrings. We will add in those to clarify what each function does.
