### 2021-09-08 825aa64
working on migrating testing to git actions instead of travis. modified some tests in the process
