### 2021-09-09 ede7b98
added more python 3 versions to gitaction pytest
