### 2021-09-09 bd92693
changing tests to python 3.8 and 3.9
