+++ .github/workflows/pythontest.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6 3.7 3.8 3.9]
+        python-version: [3.8 3.9]
 
     steps:
       - uses: actions/checkout@v2
