+++ nidm/experiment/tools/nidm_linreg.py
@@ -4,7 +4,7 @@
 # *******************************************************************************************************
 # *******************************************************************************************************
 #  nidm_linreg.py
-#  License: GPL
+#  License: Apache version 2.0
 # *******************************************************************************************************
 # *******************************************************************************************************
 # Date: 5-19-21                 Coded by: Ashmita Kumar (ashmita.kumar@gmail.com)
