+++ nidm/experiment/tools/nidm_linreg.py
@@ -4,7 +4,7 @@
 # *******************************************************************************************************
 # *******************************************************************************************************
 #  nidm_linreg.py
-#  License: Apache version 2.0
+#  License: Apache Version 2.0
 # *******************************************************************************************************
 # *******************************************************************************************************
 # Date: 5-19-21                 Coded by: Ashmita Kumar (ashmita.kumar@gmail.com)
