### 2021-09-17 137a5db
changed license in headers to be consistent with github repo Apache License Version 2.0
