+++ nidm/experiment/tools/repronim_simple2_brainvolumes.py
@@ -2,7 +2,7 @@
 #**************************************************************************************
 #**************************************************************************************
 #  repronim_simple2_brainvolumes.py
-#  License: GPL
+#  License: Apache License, Version 2.0
 #**************************************************************************************
 #**************************************************************************************
 # Date: 03-22-19                 Coded by: David Keator (dbkeator@gmail.com)
