+++ nidm/experiment/tools/bidsmri2nidm.py
@@ -2,7 +2,7 @@
 # **************************************************************************************
 # **************************************************************************************
 #   bidsmri2nidm.py
-#   License: GPL
+#   License: Apache License, Version 2.0
 # **************************************************************************************
 # **************************************************************************************
 #  Date: 10-2-17                 Coded by: David Keator (dbkeator@gmail.com)
