+++ nidm/experiment/tools/nidm_query.py
@@ -3,7 +3,7 @@
 #**************************************************************************************
 #**************************************************************************************
 #  nidm_query.py
-#  License: GPL
+#  License: Apache License, Version 2.0
 #**************************************************************************************
 #**************************************************************************************
 # Date: 8-1-18                 Coded by: David Keator (dbkeator@gmail.com)
