+++ nidm/experiment/tools/nidm_convert.py
@@ -2,7 +2,7 @@
 #**************************************************************************************
 #**************************************************************************************
 #  nidm_utils.py
-#  License: GPL
+#  License: Apache License, Version 2.0
 #**************************************************************************************
 #**************************************************************************************
 # Date: 11-28-18                 Coded by: David Keator (dbkeator@gmail.com)
