### 2021-09-19 #3578 @AKUMAR0019
Hi @osorensen. The paper, the docstrings, and the test case are expected to be added by Tuesday. We will try to have the global variable changes done the end of next weekend. However, I will inform you if they take longer.
