### 2021-09-20 a09337e
modified nidm_linreg test cases
