+++ nidm/experiment/tools/tests/test_nidm_lingreg.py
@@ -10,6 +10,8 @@ from nidm.core import Constants
 from nidm.experiment.tools.rest import RestParser
 from nidm.experiment.tools.tests.test_rest_statistics import BRAIN_VOL_FILES
 
+from nidm.experiment.tools.nidm_linreg import linear_regression
+
 import os
 from os.path import join,sep
 from pathlib import Path
@@ -18,7 +20,8 @@ import json
 from io import TextIOWrapper, BytesIO
 import subprocess
 from subprocess import PIPE
-
+import tempfile
+import click
 
 @pytest.fixture(scope="module", autouse="True")
 def setup():
@@ -39,21 +42,75 @@ def setup():
             "caltech.nidm.ttl"
         )
 
+def call_click_command(cmd, *args, **kwargs):
+    """ Wrapper to call a click command
+
+    :param cmd: click cli command function to call
+    :param args: arguments to pass to the function
+    :param kwargs: keywrod arguments to pass to the function
+    :return: None
+    """
+
+    # Get positional arguments from args
+    arg_values = {c.name: a for a, c in zip(args, cmd.params)}
+    args_needed = {c.name: c for c in cmd.params
+                   if c.name not in arg_values}
+
+    # build and check opts list from kwargs
+    opts = {a.name: a for a in cmd.params if isinstance(a, click.Option)}
+    for name in kwargs:
+        if name in opts:
+            arg_values[name] = kwargs[name]
+        else:
+            if name in args_needed:
+                arg_values[name] = kwargs[name]
+                del args_needed[name]
+            else:
+                raise click.BadParameter(
+                    "Unknown keyword argument '{}'".format(name))
+
+
+    # check positional arguments list
+    for arg in (a for a in cmd.params if isinstance(a, click.Argument)):
+        if arg.name not in arg_values:
+            raise click.BadParameter("Missing required positional"
+                                     "parameter '{}'".format(arg.name))
+
+    # build parameter lists
+    opts_list = sum(
+        [[o.opts[0], arg_values[n]] for n, o in opts.items()], [])
+    args_list = [str(v) for n, v in arg_values.items() if n not in opts]
+
+    # call the command
+    try:
+        cmd(opts_list + args_list)
+    except:
+        pass
+
+
+
 def test_simple_model():
 
 
     # run linear regression tool with simple model and evaluate output
     dirname = os.path.dirname(__file__)
     linreg_dirname = join(sep+join(*(dirname.split(sep)[:-1])))
-    cmd = ["python", join(linreg_dirname,"nidm_linreg.py"),"-nl",",".join(BRAIN_VOL_FILES),"-model",
-           'fs_000008 = DX_GROUP + http://uri.interlex.org/ilx_0100400']
 
-    subprocess.list2cmdline(cmd)
 
-    result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL)
-    out = result.stdout.decode('utf-8')
+    arguments = (dict(nidm_file_list =",".join(BRAIN_VOL_FILES),
+                      ml='fs_000008 = DX_GROUP + http://uri.interlex.org/ilx_0100400',
+                      ctr=None,regularization=None,output_file="output.txt"))
 
-    # print(out)
+    call_click_command(linear_regression,*arguments,**arguments)
+   
+
+    if os.path.exists("output.txt"):
+        fp = open("output.txt", "r")
+        out = fp.read()
+        fp.close()
+        os.remove("output.txt")
+
+
 
[…]
