### 2021-09-21 f85fe2e
upgrading map_variables_to_terms and csv2nidm to accomodate new reproschema-aligned json mapping files while maintaining backwards compatibility with previous variants of json mapping file...while also supporting BIDS json sidecar file
