+++ nidm/experiment/Utils.py
@@ -952,7 +952,43 @@ def redcap_datadictionary_to_json(redcap_dd_file,assessment_name):
 
     return json_map
 
+def detect_json_format(json_map):
+    '''
+    This function will take a json "sidecar" file or json annotation data dictionary structure
+    and determine if it''s consistent with the ReproSchema structure (compound keys root-level keys
+    DD(source=XXX,variable=YYY) and 'responseOptions' subkeys), the older pynidm format (compound
+    keys as ReproSchema, no reponseOptions subkeys), or the BIDS sidecar file structure
+    (flat structure, variable names as keys, no response options).  It will return a string associated
+    with the structure: BIDS | OLD_PYNIDM | REPROSCHEMA
 
+    :param json_map: json annotation file dictionary (file already loaded with json.load)
+
+    '''
+
+    for key,value in json_map.keys():
+        if "DD(" in key:
+            if "responseOptions" in value.keys():
+                return "REPROSCHEMA"
+            else:
+                return "OLD_PYNIDM"
+        else:
+            return "BIDS"
+
+def match_participant_id_field(source_variable):
+    '''
+    This function will test whether the source_variable is a participant ID field or not by string matching.
+    :param source_variable: source variable string to test
+    '''
+
+    if ((("participant_id" in source_variable.lower()) or
+          ("subject_id" in source_variable.lower()) or
+          (("participant" in source_variable.lower()) and ("id" in source_variable.lower())) or
+          (("subject" in source_variable.lower()) and ("id" in source_variable.lower())) or
+          (("sub" in source_variable.lower()) and ("id" in source_variable.lower())))):
+
+        return True
+    else:
+        return False
 
 def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_source=None,bids=False,owl_file='nidm',
                            associate_concepts=True, dataset_identifier=None):
@@ -972,7 +1008,7 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
     # dictionary mapping column name to preferred term
     column_to_terms = {}
 
-    # check if user supplied a JSON file and we already know a mapping for this column
+    # check if user supplied a JSON file or a json dictionary
     if json_source is not None:
         try:
             # check if json_source is a file
@@ -1034,7 +1070,8 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
             #try:
                 # check for column in json file
             try:
-                json_key = [key for key in json_map if column.lstrip().rstrip() == key.split("variable")[1].split("=")[1].split(")")[0].lstrip("'").rstrip("'")]
+                json_key = [key for key in json_map if column.lstrip().rstrip() ==
+                        key.split("variable")[1].split("=")[1].split(")")[0].lstrip("'").rstrip("'")]
             except Exception as e:
                 if "list index out of range" in str(e):
                     json_key = [key for key in json_map if column.lstrip().rstrip() == key]
@@ -1076,38 +1113,9 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
                         column_to_terms[current_tuple]['url'] = json_map[json_key[0]]['url']
                         print("url: %s" %column_to_terms[current_tuple]['url'])
                     # print("Variable: %s" %column_to_terms[current_tuple]['variable'])
-
-                    if 'levels' in json_map[json_key[0]]:
-                        column_to_terms[current_tuple]['levels'] = json_map[json_key[0]]['levels']
-                        print("levels: %s" %column_to_terms[current_tuple]['levels'])
-                    elif 'Levels' in json_map[json_key[0]]:
-                        column_to_terms[current_tuple]['levels'] = json_map[json_key[0]]['Levels']
[…]
