### 2021-09-22 aad804c
removed warnings from gitannex checks for links to files/datasets
