### 2021-09-22 374d2fd
changed gitaction for testing to be on pull request instead of push
