### 2021-09-22 755347e
fixed bug in map_vars_to_terms where on specific use cases isAbout was a dictionary instead of a list for the participant id field
