+++ nidm/experiment/Utils.py
@@ -1064,7 +1064,6 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
 
         # set up a dictionary entry for this column
         current_tuple = str(DD(source=assessment_name, variable=column))
-        column_to_terms[current_tuple] = {}
 
         # if we loaded a json file with existing mappings
         try:
@@ -1077,12 +1076,12 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
             except Exception as e:
                 if "list index out of range" in str(e):
                     json_key = [key for key in json_map if column.lstrip().rstrip() == key]
-                    #for key in json_map:
-                    #    print("%s == %s?" %(key,column.lstrip().rstrip()))
+
 
             finally:
 
                 if (json_map is not None) and (len(json_key)>0):
+                    column_to_terms[current_tuple] = {}
 
                     # added in case for some reason there isn't a label key, try source_variable and if it's
                     # a key then add this as the label as well.
@@ -1298,7 +1297,7 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
                             json_map[json_key[0]]['hasUnit']
                         print("unitCode: %s" % column_to_terms[current_tuple]['responseOptions']['unitCode'])
 
-
+                    print("***************************************************************************************")
                     print("---------------------------------------------------------------------------------------")
 
             if (json_map is not None) and (len(json_key)>0):
@@ -1323,10 +1322,12 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
             column_to_terms[subjid_tuple]['label'] = search_term
             column_to_terms[subjid_tuple]['description'] = "subject/participant identifier"
             column_to_terms[subjid_tuple]['source_variable'] = str(search_term)
-            column_to_terms[subjid_tuple]['valueType'] = URIRef(Constants.XSD["string"])
-            column_to_terms[subjid_tuple]['isAbout'] = {}
-            column_to_terms[subjid_tuple]['isAbout']['@id'] = Constants.NIDM_SUBJECTID.uri
-            column_to_terms[subjid_tuple]['isAbout']['label'] = Constants.NIDM_SUBJECTID.localpart
+            # added to support reproschema format
+            column_to_terms[subjid_tuple]['responseOptions'] = {}
+            column_to_terms[subjid_tuple]['responseOptions']['valueType'] = URIRef(Constants.XSD["string"])
+            column_to_terms[subjid_tuple]['isAbout'] = []
+            column_to_terms[subjid_tuple]['isAbout'].append({'@id':Constants.NIDM_SUBJECTID.uri,
+                        'label':Constants.NIDM_SUBJECTID.localpart})
             # column_to_terms[subjid_tuple]['variable'] = str(column)
 
             print("Variable %s automatically mapped to participant/subject identifier" %search_term)
@@ -1338,6 +1339,9 @@ def map_variables_to_terms(df,directory, assessment_name, output_file=None,json_
             continue
         # if we haven't already found an annotation for this column then have user create one.
         if current_tuple not in column_to_terms.keys():
+            # create empty annotation structure for this source variable
+            column_to_terms[current_tuple] = {}
+            # enter user interaction function to get data dictionary annotations from user
             annotate_data_element(column, current_tuple, column_to_terms)
         # then ask user to find a concept if they selected to do so
         if associate_concepts:
@@ -1403,10 +1407,23 @@ def write_json_mapping_file(source_variable_annotations, output_file, bids=False
     if bids:
         # convert to simple keys
         temp_dict = tupleKeysToSimpleKeys(source_variable_annotations)
+
+        new_dict = {}
[…]
