### 2021-09-23 47ec742
added test function for map_vars_to_terms function that does the annotations and writes the JSON mapping files. also increased version number
