+++ nidm/version.py
@@ -3,8 +3,8 @@ import os.path
 
 # Format expected by setup.py and doc/source/conf.py: string of form "X.Y.Z"
 _version_major = 3
-_version_minor = 8 
-_version_micro = '6'  # use '' for first of series, number for 1 and above
+_version_minor = 9 
+_version_micro = '0'  # use '' for first of series, number for 1 and above
 _version_extra = ''
 # _version_extra = ''  # Uncomment this for full releases
 
