### 2021-09-23 bead252
added both push and pull_request to python test gitaction
