+++ .github/workflows/pythontest.yml
@@ -1,6 +1,6 @@
 name: PyNIDM Testing
 
-on: [pull_request]
+on: [push,pull_request]
 
 jobs:
   build:
