+++ .github/workflows/pythontest.yml
@@ -1,6 +1,6 @@
 name: PyNIDM Testing
 
-on: [push,pull_request]
+on: [pull_request]
 
 jobs:
   build:
