### 2021-09-23 6817010
fixing up readthedocs text
