### 2021-09-23 dcc4129
temporarily removed details about pynidm linear-regression as formatting was messed up.
