+++ nidm/version.py
@@ -4,7 +4,7 @@ import os.path
 # Format expected by setup.py and doc/source/conf.py: string of form "X.Y.Z"
 _version_major = 3
 _version_minor = 9 
-_version_micro = '0'  # use '' for first of series, number for 1 and above
+_version_micro = '1'  # use '' for first of series, number for 1 and above
 _version_extra = ''
 # _version_extra = ''  # Uncomment this for full releases
 
