+++ docs/source/index.rst
@@ -286,470 +286,6 @@ Options:
                                   of query
   --help                          Show this message and exit.
 
-To use the linear regression algorithm successfully, structure, syntax, and querying is important. Here is how to maximize the usefulness of the tool:
-
-
-First, use pynidm query to discover the variables to use. PyNIDM allows for the use of either data elements (PIQ_tca9ck), specific URLs (http://uri.interlex.org/ilx_0100400), or source variables (DX_GROUP).
-
-An example of a potential query is: pynidm query -nl /Users/Ashu/Downloads/simple2_NIDM_examples/datasets.datalad.org/abide/RawDataBIDS/CMU_a/nidm.ttl,/Users/Ashu/Downloads/simple2_NIDM_examples/datasets.datalad.org/abide/RawDataBIDS/CMU_b/nidm.ttl -u /projects?fields=fs_000008,DX_GROUP,PIQ_tca9ck,http://uri.interlex.org/ilx_0100400
-
-You can also do:
-pynidm query -nl /simple2_NIDM_examples/datasets.datalad.org/abide/RawDataBIDS/CMU_a/nidm.ttl,/Users/Ashu/Downloads/simple2_NIDM_examples/datasets.datalad.org/abide/RawDataBIDS/CMU_b/nidm.ttl -gf fs_000008,DX_GROUP,PIQ_tca9ck,http://uri.interlex.org/ilx_0100400
-
-The query looks in the two files specified in the -nl parameter for the variables specified. In this case, we use fs_000008 and DX_GROUP (source variables), a URL (http://uri.interlex.org/ilx_0100400), and a data element (PIQ_tca9ck). The output of the file is slightly different depending on whether you use -gf or -u. With -gf, it will return the variables from both files separately, while -u combines them.
-
-For -gf, the file output is:
-
-``
-subject                               label                                    value  unit    isAbout
-------------------------------------  ----------------------------  ----------------  ------  ----------------------------------------
-ea9510e8-1561-11eb-b5e0-1094bbf2086c  PIQ                              104                    http://uri.interlex.org/base/ilx_0739363
-ea9510e8-1561-11eb-b5e0-1094bbf2086c  diagnostic group                   2            NA
-ea9510e8-1561-11eb-b5e0-1094bbf2086c  age at scan                       33            years   http://uri.interlex.org/ilx_0100400
-ea9510e8-1561-11eb-b5e0-1094bbf2086c  Supratentorial volume (mm^3)  977207            mm^3
-e940d16e-1561-11eb-b5e0-1094bbf2086c  Supratentorial volume (mm^3)       1.1513e+06   mm^3
-e940d16e-1561-11eb-b5e0-1094bbf2086c  age at scan                       21            years   http://uri.interlex.org/ilx_0100400
-e940d16e-1561-11eb-b5e0-1094bbf2086c  diagnostic group                   2            NA
-e940d16e-1561-11eb-b5e0-1094bbf2086c  PIQ                              108                    http://uri.interlex.org/base/ilx_0739363
-e9ead2c2-1561-11eb-b5e0-1094bbf2086c  Supratentorial volume (mm^3)       1.01447e+06  mm^3
-e9ead2c2-1561-11eb-b5e0-1094bbf2086c  age at scan                       21            years   http://uri.interlex.org/ilx_0100400
-e9ead2c2-1561-11eb-b5e0-1094bbf2086c  diagnostic group                   2            NA
[…]
