### 2021-09-29 #3578 @AKUMAR0019
Hi @osorensen. The test case has been updated (https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/tests/test_nidm_lingreg.py), and we have addressed all the points raised by the reviewers except for the feedback on global variables, which is taking a while to implement as tests and other code is already functioning with those variables. However, the code itself is still working, the test specific to the code is running properly, the paper has been updated, and the license and references have been fixed, so I think we are ready to proceed.
