### 2021-10-05 #3578 @osorensen
@AKUMAR0019, it looks like there are no papers to compile in the source repository. Can you please make sure the paper is there? If you wish, it is possible to put the paper in a separate branch, and then compile it with `whedon generate pdf from branch branch-name`.
