### 2021-10-05 #3578 @AKUMAR0019
@osorensen @robbisg While https://github.com/incf-nidash/PyNIDM is where the code is stored, the paper is in https://github.com/AKUMAR0019/PyNIDM/blob/master/paper.md. My fellow contributors to PyNIDM wanted me to keep the paper in my personal fork of it as opposed to putting it in the main repository. I am not certain how to get whedon to compile the paper there. Can you advise? Thank you.
