### 2021-10-05 #3578 @osorensen
@AKUMAR0019, the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) say:

> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).

I hence suggest that you discuss this again with your fellow contributors and see if you can add the paper to the repository.
