### 2021-10-05 #3578 @danielskatz
Note that the paper was in the repo when this submission was accepted for review, and it needs to stay there, though it can be a branch, as @osorensen says.  Otherwise, we will have to reject the submission.
