### 2021-10-05 #3578 @osorensen
Thanks @danielskatz. I just wanted to add that the original submission repository was the fork where the paper is saved. Thus the paper compilation succeeded in the pre review issue. However, since this fork is not the home of the software, we changed the submission repository after some discussion in the top of this thread. And since the paper has not been put in this repository, compilation now fails.
