### 2021-10-07 #3578 @AKUMAR0019
@osorensen @robbisg @htwangtw Thank you for your patience. The paper is now available at https://github.com/incf-nidash/PyNIDM/tree/master/docs/manuscripts/linreg_joss, and it is compiling.
