### 2021-10-12 f6bdb1a
updated nidm_linreg.py tool so output file is written correctly. updated version minor number
