### 2021-10-13 #3578 @AKUMAR0019
Hi @osorensen @robbisg @htwangtw. Thank you for your time. Just to recap: The repository listed by whedon (https://github.com/incf-nidash/PyNIDM) contains the source code at https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/nidm_linreg.py. The plain-text license file is also in the main repo (https://github.com/incf-nidash/PyNIDM/blob/master/LICENSE). The code itself contains documentation, as does the README.

There are automated tests running for both the RestParser and for the linear regression tool, which can be found at https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/tests/test_rest.py and https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/tests/test_nidm_lingreg.py. The README contains a button called "docs passing" that has redirects to the documentation (also found at https://pynidm.readthedocs.io/en/latest/). There is a section there (1.5-1.7) explaining how to contribute to the software, report issues, and seek support.

The paper is fully updated with references as described in the link in the checklist, and it has changed based on your feedback. It can be found at https://github.com/incf-nidash/PyNIDM/tree/master/docs/manuscripts/linreg_joss.

Is there anything else I need to do for the review?
