### 2021-10-13 aa49ed1
added additional readme details about linear regression tool
