### 2021-10-13 f18bee7
added additional readme details about linear regression tool
