### 2021-10-18 a72f4e1
updated Installation section of README
