### 2021-10-18 f7097c2
changed PyPi to Installation in the ReadTheDocs documentation
