### 2021-10-19 #3578 @AKUMAR0019
> @AKUMAR0019 @robbisg @htwangtw; after giving it some thought, I've decided to add a `query-scope` label on this submission. I share the concern stated by @robbisg and @htwangtw in their posts just above that the contribution of this submission might be too small to be of sufficient scholarly effort for a JOSS publication. In particular, I am not sure if the submission satisfies the following criteria:
>
> > “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.
>
> @openjournals/joss-eics; please note that the submission is an extension of an existing tool, and is mainly [this file](https://github.com/incf-nidash/PyNIDM/blob/5d7e502a1edf523ff8173127e9782db3b3197692/nidm/experiment/tools/nidm_linreg.py), if I have understood correctly.

@osorensen The linear regression tool uses the REST-API to query NIDM files which is located through this link (https://github.com/incf-nidash/PyNIDM/blob/master/nidm/experiment/tools/rest.py) and should be added to the lines of code in the linear regression tool to get an accurate estimate.  In our interpretation of the JOSS guidelines, the total lines of code exceeds the journal minimum.  Further, one can find the PyNIDM code included in the source file for additional reference to code written as part of this submission: “from nidm.experiment.Query import GetProjectsUUID”, “from nidm.experiment.tools.rest import RestParser”.

The scholarly effort is the development of a tool capable of learning linear models in the context of linked-data, RDF documents for human neuroimaging-related neuroscience (i.e. NIDM).  There are no other tools supporting such statistical investigation of human neuroimaging-related neuroscience data/metadata stored in RDF documents that we know of.  The Neuroimaging Data Model (NIDM) was started by an international team of cognitive scientists, computer scientists and statisticians to develop a data format capable of describing all aspects of the data lifecycle, from raw data through analyses and provenance.  The NIDM working group officially began meeting in 2011 focusing on improving how data, analyses, and provenance are shared in the neuroimaging community. The NIDM group participates in monthly teleconference meetings, with 9-12 attendees per month on average, representing 12 institutions across Europe and the US.  Over eighteen abstracts have been published at international meetings as well as 3 journal publications describing the group’s work. All NIDM group meeting minutes and twenty code repositories are publicly available on GoogleDocs and GitHub, respectively, with a total of over 215 meeting records, demonstrating the group's commitment to open, community-based, development and evidencing the scholarly effort of this work.

This work is not an extension of an existing tool.  It is a new tool built on top of the NIDM work which spans a wide range of functionality to track the lifecycle of data from acquisition to results.  This particular tool was developed along with the associated query functionality, linked above, to make running linear models on linked-data NIDM graphs possible across different datasets.  There is no other tool that  provides such functionality.
