### 2021-11-04 #3578 @AKUMAR0019
@arfon @osorensen @robbisg @htwangtw We are sorry to hear of JOSS’s final decision. Upon initial review these concerns should have been brought up and an initial decision (minor revisions, major revisions, or rejection) made.  Instead we received ad-hoc comments and requests for changes which we made. After a couple of rounds this paper was ultimately rejected. This process is unlike any scientific journal we have published in and is a process we strongly encourage JOSS to change as it wastes both ours and your time. If the original paper did not meet the JOSS guidelines then it should have been rejected straight away and we would have used our valuable time to submit it to another journal.  Please consider revising your review methodology.
