### 2021-11-04 #3578 @arfon
@AKUMAR0019 – I'm sorry that you are disappointed in the review experience here. JOSS reviews _by design_ are interactive, conversational, and iterative. We consider this a **feature** of our review experience, and it has stood us in good stead over the past 5 years and nearly 1500 published papers.

This submission was a little more complex than some for us as your tool is a small part of a much larger piece of software. As such, the concern around the "Substantial scholarly effort" criterion only became apparent well into the review process which while later than most submissions, is still a valid time for this to be raised. Overall, I believe the reviewers and editor performed their duties very well here and I thank them for their time.
